* chunker.c: allow chunk to be written on different holding disk. * server-src/driver.c: allow chunk to be written on different   holding disk. * server-src/conffile.c: make sure that disksize and chunksize   are multiple of TAPE_BLOCK_SIZE. * server-src/conffile.h: disksize and chunksize are long. * server-src/diskfile.c (find_disk): new function to see if a   given disk is contained in a list structure. * server-src/diskfile.h (find_disk): prototype. * server-src/driverio.h (RQ_MORE_DISK): Added message. * server-src/driverio.h (free_assignedhd): prototype. * server-src/driverio.h (assignedhd_t): new data structure. * server-src/driverio.c (cmdstr): added comamnd RQ-MORE-DISK. * server-src/driverio.c (free_assignedhd): new function to free an   array of assignedhd_t.

+++ server-src/conffile.h
@@ -25,7 +25,7 @@
  *			   University of Maryland at College Park
  */
 /*
- * $Id: conffile.h,v 1.35 2000/04/09 07:59:04 oliva Exp $
+ * $Id: conffile.h,v 1.36 2000/05/27 22:45:28 martinea Exp $
  *
  * interface for config file reading code
  */
@@ -162,8 +162,8 @@ typedef struct holdingdisk_s {
 
     char *comment;
     char *diskdir;
-    int disksize;
-    int chunksize;
+    long disksize;
+    long chunksize;
 
     int s_comment;
     int s_disk;
